requirements.txt
asgiref==3.8.1
blinker==1.9.0
boto3==1.38.14
botocore==1.38.14
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.4.1
django-crispy-forms==2.3
django-cors-headers==4.6.0
django-storages==1.14.6
djangorestframework==3.15.2
django-test-plus==2.2.4
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
rich==13.9.4
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
whitenoise
cloudinary
django-cloudinary-storage
